requirements.txt
web3>=6.11.0
eth-account>=0.9.0
python-dotenv>=1.0.0
solcx>=2.0.0
